services/meal-planner-service/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# OpenAI
openai==1.3.7

# Authentication & Security
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client (for auth service communication)
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

